+++ packages/entities/entities-plugins/sandbox/pages/PluginFormPage.vue
@@ -69,6 +69,8 @@ provide(FEATURE_FLAGS.KM_2306_CONDITION_FIELD_314, true)
 
 useProvideExperimentalFreeForms([
   'service-protection',
+  'prometheus',
+  'aws-lambda',
 ])
 
 const konnectConfig = ref<KonnectPluginFormConfig>({