+++ packages/entities/entities-plugins/src/components/free-form/shared/composables/labels.ts
@@ -32,6 +32,11 @@ const labelDictionary: Record<string, string> = {
   azure: 'Azure',
   acl: 'ACL',
   cookie: 'Cookie',
+  ai: 'AI',
+  awsgateway: 'AWS Gateway',
+  imds: 'IMDS',
+  sts: 'STS',
+  arn: 'ARN',
 }
 
 export function replaceByDictionary(name: string) {