+++ packages/entities/entities-plugins/src/utils/free-form.ts
@@ -38,6 +38,14 @@ const mapping = {
     experimental: true,
     component: 'CommonForm',
   },
+  'prometheus': {
+    experimental: true,
+    component: 'CommonForm',
+  },
+  'aws-lambda': {
+    experimental: true,
+    component: 'CommonForm',
+  },
 } as const
 
 export type FreeFormName = keyof typeof mapping