+++ packages/analytics/analytics-config-store/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@1.3.3...@kong-ui-public/analytics-config-store@1.3.4) (2026-03-23)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-config-store
+
+
+
+
+
+## [1.3.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@1.3.2...@kong-ui-public/analytics-config-store@1.3.3) (2026-03-23)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-config-store
+
+
+
+
+
+## [1.3.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@1.3.1...@kong-ui-public/analytics-config-store@1.3.2) (2026-03-20)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-config-store
+
+
+
+
+
 ## [1.3.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@1.3.0...@kong-ui-public/analytics-config-store@1.3.1) (2026-03-18)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-config-store