+++ packages/core/split-pane/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.8](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/split-pane@0.3.7...@kong-ui-public/split-pane@0.3.8) (2026-03-21)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @kong/icons to ^1.47.0 ([#3032](https://github.com/Kong/public-ui-components/issues/3032)) ([8f04a91](https://github.com/Kong/public-ui-components/commit/8f04a919b363b9ed27d6b7b8b890f679fb9f604b))
+
+
+
+
+
 ## [0.3.7](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/split-pane@0.3.6...@kong-ui-public/split-pane@0.3.7) (2026-03-18)